Created a SearchInput component, but the logic doesn't work yet
